start hashed refactor work
